fixed grid generation

+++ collect-server/src/main/java/org/openforis/collect/manager/RandomValuesGenerator.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public abstract class RandomValuesGenerator {
 
-	public static <T> List<T> generateRandomSubset(List<T> collection, float percentageOverTotal) {
+	public static <T> List<T> generateRandomSubset(List<T> collection, double percentageOverTotal) {
 		List<T> result = new ArrayList<T>();
 		int numberOfRecordsToGenerate = (int) Math.ceil(((collection.size() * percentageOverTotal) / 100));
 		List<T> currentCollection = new ArrayList<T>(collection);
